+++ src/QuickFixProvider.ts
@@ -10,9 +10,9 @@ import {
   Diagnostic,
   CodeActionKind,
   WorkspaceEdit,
-  Position
-} from "vscode";
-import { SOURCE } from "./Linter";
+  Position,
+} from 'vscode';
+import { SOURCE } from './Linter';
 
 export class QuickFixProvider implements CodeActionProvider {
   provideCodeActions(
@@ -22,28 +22,60 @@ export class QuickFixProvider implements CodeActionProvider {
     token: CancellationToken
   ): CodeAction[] {
     const warnings = context.diagnostics.filter(
-      diagnostic =>
+      (diagnostic) =>
         diagnostic.source === SOURCE &&
         diagnostic.severity === DiagnosticSeverity.Warning
     );
-    return warnings.length > 0 ? [this.createFix(document, warnings)] : [];
+    return warnings.length > 0 ? this.createActions(document, warnings) : [];
   }
 
-  private createFix(
+  private createActions(
     document: TextDocument,
     diagnostics: Diagnostic[]
+  ): CodeAction[] {
+    return diagnostics.map((diagnostic) => {
+      return diagnostic.code === 'RuboCop'
+        ? this.createRubocopAction(document, diagnostic)
+        : this.createHamlLintAction(document, diagnostic);
+    });
+  }
+
+  private createHamlLintAction(
+    document: TextDocument,
+    diagnostic: Diagnostic
   ): CodeAction {
-    const linterNames = diagnostics
-      .map(({ code }) => code)
-      // TODO: Support for RuboCop
-      .filter((code) => code !== "RuboCop")
-      .join(", ");
-    const fix = new CodeAction("Disable linters", CodeActionKind.QuickFix);
+    const fix = new CodeAction(
+      `Disable ${diagnostic.code} for this entire file`,
+      CodeActionKind.QuickFix
+    );
     fix.edit = new WorkspaceEdit();
     fix.edit.insert(
       document.uri,
       new Position(0, 0),
-      `-# haml-lint:disable ${linterNames}\n`
+      `-# haml-lint:disable ${diagnostic.code}\n`
+    );
+    return fix;
+  }
+
+  private createRubocopAction(
+    document: TextDocument,
+    diagnostic: Diagnostic
+  ): CodeAction {
+    const rule = diagnostic.message.split(':')[0];
+    const fix = new CodeAction(
+      `Disable ${rule} for this entire file`,
+      CodeActionKind.QuickFix
+    );
+    fix.edit = new WorkspaceEdit();
+    fix.edit.insert(
+      document.uri,
+      new Position(0, 0),
+      `-# rubocop:disable ${rule}\n`
+    );
+    fix.edit.insert(
+      document.uri,
+      new Position(document.lineCount + 1, 0),
+      `-# rubocop:enable ${rule}\n`
     );
     return fix;
   }